[core] Move task state test to sm package
